создает рабочую версию презентера список точек маршрута
