+++ src/presenter/route-list.js
@@ -3,37 +3,21 @@ import SortView from '../view/sort';
 import TripPointView from '../view/trip-point';
 import TripPointFormView from '../view/trip-point-form';
 import ListEmptyView from '../view/list-empty';
-import TripCitiesView from '../view/trip-cities';
-import TripPriceView from '../view/trip-price';
-import PointPresenter from 'point';
-import {render} from '../utils/render';
+import {render, replace} from '../utils/render';
 
-export default class routeList {
+export default class RouteList {
   constructor(routeListContainer) {
     this._routeListContainer = routeListContainer;
-    this._pointPresenter = new Map();
 
     this._tripPointsListCopmonent = new TripPointsListView();
     this._sortComponent = new SortView();
     this._emptyListComponent = new ListEmptyView();
-
-    this._onePointChange = this._onePointChange.bind(this);
-    this._onModeChange = this._onModeChange.bind(this);
-
-    this._tripInfoCitiesComponent = new TripCitiesView();
-    this._tripPriceComponent = new TripPriceView();
   }
 
   init(points) {
     this._points = points.slice();
-
     render(this._routeListContainer, this._tripPointsListCopmonent);
-
     this._renderRouteList();
-  }
-
-  _onModeChange() {
-    this._pointPresenter.forEach((presenter) => presenter.resetView());
   }
 
   _renderSort() {
@@ -41,9 +25,40 @@ export default class routeList {
   }
 
   _renderPoint(point) {
-    const pointPresenter = new PointPresenter(this._tripPointsListCopmonent, this._onPointChange, this._onModeChange);
-    pointPresenter.init(point);
-    this._pointPresenter.set(point.id, pointPresenter);
+    const pointComponent = new TripPointView(point);
+    const pointFormComponent = new TripPointFormView(point);
+
+    const replacePointViewToForm = () => {
+      replace(pointFormComponent, pointComponent);
+    };
+
+    const replaceFormToPointView = () => {
+      replace(pointComponent, pointFormComponent);
+    };
+
+    const onEscKeyDown = (evt) => {
+      if (evt.key === 'Escape' || evt.key === 'Esc') {
+        evt.preventDefault();
+        replaceFormToPointView();
+        document.removeEventListener('keydown', onEscKeyDown);
+      }
+    };
+
+    pointComponent.setFormClickHandler(() => {
+      replacePointViewToForm();
+      document.addEventListener('keydown', onEscKeyDown);
+    });
+
+    pointFormComponent.setFormSubmitHandler(() => {
+      replaceFormToPointView();
+      document.removeEventListener('keydown', onEscKeyDown);
+    });
+
+    pointFormComponent.setFormClickCloseHandler(() => {
+      replaceFormToPointView();
+    });
+
+    render(this._tripPointsListCopmonent, pointComponent);
   }
 
   _renderPoints () {
@@ -62,19 +77,6 @@ export default class routeList {
 
     this._renderSort();
     this._renderPoints();
-  }
-
-  _clearPointsList() {
-    this._pointPresenter.forEach((presenter) => presenter.destroy());
-    this._pointPresenter.clear();
-  }
-
-  _renderRouteCities() {
-
-  }
-
-  _renderTripPrice() {
-
   }
 }
